Who ever said this was effective
